Add linux compatible camera sensor
The existing ColorWebcam sensor relies on the Windows-only videoInput
library. However, a linux-compatible solution can be built using
cct-io's existing opencv dependency.

The included implementation has been demonstrated to work on Ubuntu
14.04 with an HP 4310 webcam, but only with its default settings of
640x480 @ 20 fps (as an MJPEG stream). The camera itself is capable of
1080p30 output (in H.264 format), so the inability to go beyond VGA
resolutions is either the result of improper use of the javacv/opencv
APIs or a limitation of the V4L(2) driver.